adding player controller to network
