+++ Assets/Prefabs/NetworkPlayer.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
 guid: f14ba83094eea4b2588f0710b74d34ff
-timeCreated: 1486007248
+timeCreated: 1486609513
 licenseType: Free
 NativeFormatImporter:
   userData:
