+++ Assets/Scripts/NetworkCameraScript.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 33659ec1465ec4d3980d7fc0cdd348fc
+timeCreated: 1486610683
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
