+++ Assets/Scripts/PlayerController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: dad1a72d5cb7e4da7838d80b76768149
+timeCreated: 1483585403
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
