+++ ProjectSettings/GraphicsSettings.asset
@@ -33,14 +33,14 @@ GraphicsSettings:
   m_SpritesDefaultMaterial: {fileID: 10754, guid: 0000000000000000f000000000000000,
     type: 0}
   m_TierSettings_Tier1:
-    renderingPath: 0
-    useCascadedShadowMaps: 0
+    renderingPath: 1
+    useCascadedShadowMaps: 1
   m_TierSettings_Tier2:
-    renderingPath: 0
-    useCascadedShadowMaps: 0
+    renderingPath: 1
+    useCascadedShadowMaps: 1
   m_TierSettings_Tier3:
-    renderingPath: 0
-    useCascadedShadowMaps: 0
+    renderingPath: 1
+    useCascadedShadowMaps: 1
   m_DefaultRenderingPath: 1
   m_DefaultMobileRenderingPath: 1
   m_TierSettings: []
