+++ build.app/Contents/Data/Managed/etc/mono/config
@@ -1,0 +1,27 @@
+<configuration>
+	<dllmap dll="i:cygwin1.dll" target="libc.dylib" os="!windows" />
+	<dllmap dll="libc" target="libc.dylib" os="!windows"/>
+	<dllmap dll="intl" target="libintl.dylib" os="!windows"/>
+	<dllmap dll="intl" name="bind_textdomain_codeset" target="libc.dylib" os="solaris"/>
+	<dllmap dll="libintl" name="bind_textdomain_codeset" target="libc.dylib" os="solaris"/>
+	<dllmap dll="libintl" target="libintl.dylib" os="!windows"/>
+	<dllmap dll="i:libxslt.dll" target="libxslt.dylib" os="!windows"/>
+	<dllmap dll="i:odbc32.dll" target="libodbc.dylib" os="!windows"/>
+	<dllmap dll="i:odbc32.dll" target="libiodbc.dylib" os="osx"/>
+	<dllmap dll="oci" target="libclntsh.dylib" os="!windows"/>
+	<dllmap dll="db2cli" target="libdb2_36.dylib" os="!windows"/>
+	<dllmap dll="MonoPosixHelper" target="libMonoPosixHelper.dylib" os="!windows" />
+	<dllmap dll="i:msvcrt" target="libc.dylib" os="!windows"/>
+	<dllmap dll="i:msvcrt.dll" target="libc.dylib" os="!windows"/>
+	<dllmap dll="sqlite" target="libsqlite.0.dylib" os="!windows"/>
+	<dllmap dll="sqlite3" target="libsqlite3.0.dylib" os="!windows"/>
+	<dllmap dll="libX11" target="libX11.dylib" os="!windows" />
+	<dllmap dll="libcairo-2.dll" target="libcairo.so.2" os="!windows"/>
+	<dllmap dll="libcups" target="libcups.so.2" os="!windows"/>
+	<dllmap dll="i:kernel32.dll">
+		<dllentry dll="__Internal" name="CopyMemory" target="mono_win32_compat_CopyMemory"/>
+		<dllentry dll="__Internal" name="FillMemory" target="mono_win32_compat_FillMemory"/>
+		<dllentry dll="__Internal" name="MoveMemory" target="mono_win32_compat_MoveMemory"/>
+		<dllentry dll="__Internal" name="ZeroMemory" target="mono_win32_compat_ZeroMemory"/>
+	</dllmap>
+</configuration>
