+++ build.app/Contents/Resources/DefaultPreferences.plist
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+<key>NSQuitAlwaysKeepsWindows</key>
+<false/>
+<key>Screenmanager Is Fullscreen mode</key>
+<string>True</string>
+</dict>
+</plist>
