+++ build.app/Contents/Resources/KeyConfig.nib/classes.nib
@@ -1,0 +1,12 @@
+{
+    IBClasses = (
+        {CLASS = FirstResponder; LANGUAGE = ObjC; SUPERCLASS = NSObject; }, 
+        {
+            CLASS = KeyConfig; 
+            LANGUAGE = ObjC; 
+            OUTLETS = {"m_KeyLabel" = id; "m_Progress" = id; }; 
+            SUPERCLASS = NSObject; 
+        }
+    ); 
+    IBVersion = 1; 
+}
