+++ build.app/Contents/Resources/ScreenSelector.nib/classes.nib
@@ -1,0 +1,57 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>IBClasses</key>
+	<array>
+		<dict>
+			<key>ACTIONS</key>
+			<dict>
+				<key>Play</key>
+				<string>id</string>
+				<key>Quit</key>
+				<string>id</string>
+				<key>SaveAndQuit</key>
+				<string>id</string>
+			</dict>
+			<key>CLASS</key>
+			<string>ScreenSelector</string>
+			<key>LANGUAGE</key>
+			<string>ObjC</string>
+			<key>OUTLETS</key>
+			<dict>
+				<key>m_GraphicsQuality</key>
+				<string>id</string>
+				<key>m_Input</key>
+				<string>id</string>
+				<key>m_OptionKey</key>
+				<string>id</string>
+				<key>m_ScreenResolution</key>
+				<string>id</string>
+				<key>m_Windowed</key>
+				<string>id</string>
+			</dict>
+			<key>SUPERCLASS</key>
+			<string>NSObject</string>
+		</dict>
+		<dict>
+			<key>CLASS</key>
+			<string>FirstResponder</string>
+			<key>LANGUAGE</key>
+			<string>ObjC</string>
+			<key>SUPERCLASS</key>
+			<string>NSObject</string>
+		</dict>
+		<dict>
+			<key>CLASS</key>
+			<string>ScreenSelectorTableView</string>
+			<key>LANGUAGE</key>
+			<string>ObjC</string>
+			<key>SUPERCLASS</key>
+			<string>NSTableView</string>
+		</dict>
+	</array>
+	<key>IBVersion</key>
+	<string>1</string>
+</dict>
+</plist>
